Publish sortEntries -> SortEntries
