+++ pkg/sparse/entry.go
@@ -1,5 +1,7 @@
 package sparse
 
+import "sort"
+
 // Entry is an entry in a sparse vector or matrix.
 type Entry struct {
 	// Index is the index of the entry.
@@ -54,4 +56,16 @@ func (a CSCEntriesSort) Less(i, j int) bool {
 		return false
 	}
 	return false
+}
+
+type entrySort []Entry
+
+func (a entrySort) Len() int           { return len(a) }
+func (a entrySort) Less(i, j int) bool { return a[i].Index < a[j].Index }
+func (a entrySort) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+
+func SortEntries(entries []Entry) []Entry {
+	entries = append(entries[:0:0], entries...)
+	sort.Sort(entrySort(entries))
+	return entries
 }
